# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/timeoff_balances:
    get:
      tags:
        - TimeoffBalances
      summary: List TimeoffBalances
      description: Get a paginated list of TimeoffBalances with optional filtering and sorting
      operationId: listTimeoffBalances
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeoffBalanceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TimeoffBalances
      summary: Create TimeoffBalance
      description: Create a new TimeoffBalance
      operationId: createTimeoffBalance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeoffBalanceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeoffBalance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/timeoff_requests:
    get:
      tags:
        - TimeoffRequests
      summary: List TimeoffRequests
      description: Get a paginated list of TimeoffRequests with optional filtering and sorting
      operationId: listTimeoffRequests
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeoffRequestList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TimeoffRequests
      summary: Create TimeoffRequest
      description: Create a new TimeoffRequest
      operationId: createTimeoffRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeoffRequestRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeoffRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    TimeoffBalance:
      type: object
      required:
        - id
        - company_id
        - timeoff_type_id
        - employee_id
        - period
        - allocated
        - used
        - carried_over
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        timeoff_type_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        period:
          type: string
          maxLength: 10
        allocated:
          type: string
        used:
          type: string
        accrual_plan_id:
          nullable: true
          type: string
          format: uuid
        date_from:
          nullable: true
          type: string
          format: date
        date_to:
          nullable: true
          type: string
          format: date
        last_accrual_at:
          nullable: true
          type: string
          format: date-time
        carried_over:
          type: string
        expired_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateTimeoffBalanceRequest:
      type: object
      required:
        - company_id
        - timeoff_type_id
        - employee_id
        - period
        - allocated
        - used
        - carried_over
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        timeoff_type_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        period:
          type: string
          maxLength: 10
        allocated:
          type: string
        used:
          type: string
        accrual_plan_id:
          nullable: true
          type: string
          format: uuid
        date_from:
          nullable: true
          type: string
          format: date
        date_to:
          nullable: true
          type: string
          format: date
        last_accrual_at:
          nullable: true
          type: string
          format: date-time
        carried_over:
          type: string
        expired_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    TimeoffBalanceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TimeoffBalance'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    TimeoffRequest:
      type: object
      required:
        - id
        - company_id
        - timeoff_type_id
        - employee_id
        - date_start
        - date_end
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        timeoff_type_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        date_start:
          type: string
          format: date
        date_end:
          type: string
          format: date
        note:
          nullable: true
          type: string
          maxLength: 1000
        approval_employee_id:
          nullable: true
          type: string
          format: uuid
        approval_request_id:
          nullable: true
          type: string
          format: uuid
        note_reject:
          nullable: true
          type: string
          maxLength: 1000
        status:
          type: string
        metadata:
          type: object
...
    CreateTimeoffRequestRequest:
      type: object
      required:
        - company_id
        - timeoff_type_id
        - employee_id
        - date_start
        - date_end
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        timeoff_type_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        date_start:
          type: string
          format: date
        date_end:
          type: string
          format: date
        note:
          nullable: true
          type: string
          maxLength: 1000
        approval_employee_id:
          nullable: true
          type: string
          format: uuid
        approval_request_id:
          nullable: true
          type: string
          format: uuid
        note_reject:
          nullable: true
          type: string
          maxLength: 1000
        status:
          type: string
        metadata:
          type: object
...
    TimeoffRequestList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TimeoffRequest'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []